Meal servlet now corresponds to Service class

+++ src/main/java/ru/javawebinar/topjava/service/MealService.java
@@ -4,34 +4,41 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.repository.UserRepository;
+import ru.javawebinar.topjava.repository.inmemory.InMemoryMealRepository;
 
 import java.util.Collection;
-import java.util.List;
+import java.util.stream.Collectors;
 
-import static ru.javawebinar.topjava.util.ValidationUtil.checkNotFound;
 import static ru.javawebinar.topjava.util.ValidationUtil.checkNotFoundWithId;
 
 public class MealService {
 
-    private MealRepository repository;
+    private MealRepository repository = new InMemoryMealRepository();
 
     public Meal create(Meal meal) {
         return repository.save(meal);
     }
 
-    public void delete(int id) {
+    public void delete(int id, int userId) {
+        if(repository.get(id).getUserId() != userId) return;
+
         checkNotFoundWithId(repository.delete(id), id);
     }
 
-    public Meal get(int id) {
-        return checkNotFoundWithId(repository.get(id), id);
+    public Meal get(int id, int userId) {
+        Meal meal = checkNotFoundWithId(repository.get(id), id);
+        if (meal.getUserId() != userId) return null;
+        return meal;
     }
 
-    public Collection<Meal> getAll() {
-        return repository.getAll();
+    public Collection<Meal> getAll(int userId) {
+        return repository.getAll().
+                stream().filter(meal -> meal.getUserId()==userId)
+                .collect(Collectors.toList());
     }
 
-    public void update(Meal meal) {
+    public void update(Meal meal, int userId) {
+        if (meal.getUserId() != userId) return;
         checkNotFoundWithId(repository.save(meal), meal.getId());
     }
